refactor(opencode): extract auth.json logic into dedicated opencode_auth module and test file
